+++ qbraid/passes/qasm/__init__.py
@@ -19,52 +19,31 @@ Functions
 .. autosummary::
    :toctree: ../stubs/
 
-    depth
     rebase
-    unfold_qasm2
-    decompose_qasm3
-    decompose_qasm2
     insert_gate_def
-    replace_gate_name
+    replace_gate_names
     add_stdgates_include
     remove_stdgates_include
     convert_qasm_pi_to_decimal
     normalize_qasm_gate_params
-    remove_unused_gates
-    has_measurements
-    remove_measurements
-    remove_include_statements
 
 """
-from .analyze import depth, has_measurements
 from .compat import (
     add_stdgates_include,
     convert_qasm_pi_to_decimal,
     insert_gate_def,
     normalize_qasm_gate_params,
-    remove_include_statements,
-    remove_measurements,
     remove_stdgates_include,
-    replace_gate_name,
+    replace_gate_names,
 )
-from .decompose import decompose_qasm2, decompose_qasm3, rebase
-from .format import remove_unused_gates
-from .unfold import unfold_qasm2
+from .decompose import rebase
 
 __all__ = [
-    "depth",
     "rebase",
-    "unfold_qasm2",
-    "decompose_qasm2",
-    "decompose_qasm3",
     "insert_gate_def",
-    "replace_gate_name",
+    "replace_gate_names",
     "add_stdgates_include",
     "remove_stdgates_include",
     "convert_qasm_pi_to_decimal",
     "normalize_qasm_gate_params",
-    "remove_unused_gates",
-    "has_measurements",
-    "remove_measurements",
-    "remove_include_statements",
 ]