+++ qbraid/programs/gate_model/qasm2.py
@@ -14,18 +14,20 @@ Module defining OpenQasm2Program class.
 """
 from __future__ import annotations
 
-import re
+from typing import TYPE_CHECKING
 
 import numpy as np
-from openqasm3.ast import Program
-from openqasm3.parser import parse
+import pyqasm
 from qbraid_core._import import LazyLoader
 
-from qbraid.passes.qasm import depth, normalize_qasm_gate_params, rebase, remove_measurements
+from qbraid.passes.qasm import normalize_qasm_gate_params, rebase
 from qbraid.programs.exceptions import ProgramTypeError
 from qbraid.programs.typer import Qasm2String, Qasm2StringType
 
 from ._model import GateModelProgram
+
+if TYPE_CHECKING:
+    import qbraid.runtime
 
 transpiler = LazyLoader("transpiler", globals(), "qbraid.transpiler")
 
@@ -38,58 +40,50 @@ class OpenQasm2Program(GateModelProgram):
         if not isinstance(program, Qasm2String):
             raise ProgramTypeError(message=f"Expected 'str' object, got '{type(program)}'.")
 
-    def parsed(self) -> Program:
-        """Parse the program string."""
-        return parse(self._program)
-
-    def _get_bits(self, bit_type: str) -> list[tuple[str, int]]:
-        """Return the number of qubits or classical bits in the circuit.
-
-        Args:
-            bit_type: either "q" or "c" for qubits or classical bits, respectively.
-
-        Returns:
-            list[tuple[str, int]]: A list of qubits or classical bits. Each bit is
-                a tuple of (register name, index)
-        """
-        matches = re.findall(rf"{bit_type}reg (\w+)\[(\d+)\];", self.program)
-
-        result = []
-        for match in matches:
-            var_name = match[0]
-            n = int(match[1])
-            result.extend([(var_name, i) for i in range(n)])
-        return result
+        self._module = pyqasm.loads(program)
 
     @property
-    def qubits(self) -> list[tuple[str, int]]:
-        """Use regex to extract all qreg definitions from the string"""
-        return self._get_bits("q")
+    def qubits(self) -> dict[str, int]:
+        """Return the qubits acted upon by the operations in this circuit"""
+        return self._module._qubit_registers
+
+    @property
+    def module(self) -> pyqasm.Module:
+        """Return the pyqasm module."""
+        return self._module
+
+    @property
+    def num_qubits(self) -> int:
+        """Return the number of qubits in the circuit."""
+        return self._module.num_qubits
 
     @property
     def num_clbits(self) -> int:
         """Return the number of classical bits in the circuit."""
-        return len(self._get_bits("c"))
+        return self._module.num_clbits
 
     @property
     def depth(self) -> int:
         """Return the circuit depth (i.e., length of critical path)."""
-        counts = {q: 0 for q in self.qubits}
-        program = self.parsed()
-        counts = depth(program.statements, counts)
-        return max(counts.values())
+        return self._module.depth()
 
     def _unitary(self) -> np.ndarray:
         """Return the unitary of the QASM"""
-        return transpiler.transpile(self.program, "cirq").unitary()
+        raise NotImplementedError
 
-    def transform(self, device) -> None:
+    def validate(self) -> None:
+        """Validate the QASM."""
+        self._module.validate()
+
+    def transform(self, device: qbraid.runtime.QuantumDevice, **kwargs) -> None:
         """Transform program to according to device target profile."""
         if device.id == "quera_qasm_simulator":
-            self._program = remove_measurements(self.program)
+            self._module.remove_measurements()
+            self._module.unroll()
+            self.program = pyqasm.dumps(self._module)
 
         basis_gates = device.profile.get("basis_gates")
 
         if basis_gates is not None and len(basis_gates) > 0:
-            transformed_qasm = rebase(self.program, basis_gates)
+            transformed_qasm = rebase(self.program, basis_gates, **kwargs)
             self._program = normalize_qasm_gate_params(transformed_qasm)