+++ tests/passes/test_qasm3_decompose.py
@@ -23,10 +23,11 @@ from openqasm3.parser import parse
 
 from qbraid.passes.exceptions import CompilationError, QasmDecompositionError
 from qbraid.passes.qasm.compat import normalize_qasm_gate_params
-from qbraid.passes.qasm.decompose import assert_gates_in_basis, decompose_qasm3, rebase
+from qbraid.passes.qasm.decompose import assert_gates_in_basis, rebase
 from qbraid.programs.gate_model.qasm3 import OpenQasm3Program
 
 
+# move to pyqasm
 @pytest.mark.parametrize(
     "original_program, expected_program",
     [
@@ -118,6 +119,7 @@ def test_rebase_controlled_rotation_gates(original_program: str, expected_progra
     assert converted_program == expected_program
 
 
+# move to pyqasm
 def test_rebase_bad_program_raises_decomposition_error():
     """Test conversion of QASM3 program to basis gates"""
     bad_program = """
@@ -130,6 +132,7 @@ crx() q[0], q[1];
         rebase(bad_program, {"rz", "ry", "cx"})
 
 
+# move to pyqasm
 def test_rebase_bad_program_parser_error():
     """Test conversion of QASM3 program to basis gates"""
     bad_program = """
@@ -139,7 +142,7 @@ qubit[2] q;
 crx :-) q[0], q[1];
 """
     with pytest.raises(ValueError, match="Invalid OpenQASM program."):
-        rebase(bad_program, {"crx"})
+        rebase(bad_program, "any")
 
 
 @pytest.fixture
@@ -175,26 +178,23 @@ rz(-pi / 2) q[1];
 """
 
 
-def test_decompose_qasm3(qasm_crx_program: str, qasm_crx_decomposed: str):
-    """Test decomposing a QASM3 program"""
-    converted_program = decompose_qasm3(qasm_crx_program)
-    assert converted_program.strip() == qasm_crx_decomposed.strip()
-
-
-def test_rebase_raises_for_gateset_type_error(qasm_crx_program):
-    """Test rebase raises a TypeError if gateset is not a set"""
+# move to pyqasm
+def test_rebase_raises_for_basis_gates_type_error(qasm_crx_program):
+    """Test rebase raises a TypeError if basis_gates is not a set"""
     with pytest.raises(
         TypeError, match="Basis gate set must be a set of strings or a string identifier."
     ):
         rebase(qasm_crx_program, 42)
 
 
+# move to pyqasm
 def test_rebase_raises_for_bad_basis_gate_set_identifier(qasm_crx_program):
     """Test rebase raises a TypeError if gateset is not a set"""
     with pytest.raises(ValueError, match="Invalid basis gate set identifier."):
         rebase(qasm_crx_program, gateset="null")
 
 
+# move to pyqasm
 def test_assert_gates_in_basis(qasm_crx_parsed: ast.Program):
     """Test assertion that all gates are in the basis"""
     with pytest.raises(
@@ -203,18 +203,21 @@ def test_assert_gates_in_basis(qasm_crx_parsed: ast.Program):
         assert_gates_in_basis(qasm_crx_parsed, {"h"})
 
 
+# move to pyqasm
 def test_raise_value_error_on_empty_basis_set(qasm_crx_parsed: ast.Program):
     """Test raising a ValueError when the basis set is empty"""
     with pytest.raises(ValueError, match="Basis gate set cannot be empty."):
         rebase(qasm_crx_parsed, set())
 
 
+# move to pyqasm
 def test_rebase_program_already_in_basis_set(qasm_crx_program: str):
     """Test that a program already in the basis set is not modified"""
     converted_program = rebase(qasm_crx_program, {"crx"})
     assert converted_program.strip() == qasm_crx_program.strip()
 
 
+# move to pyqasm
 def test_rebase_raises_for_unsatisfied_predicates(qasm_crx_program: str):
     """Test that rebase raises an error if the predicates are not satisfied"""
     gateset = {"h"}
@@ -227,11 +230,13 @@ def test_rebase_raises_for_unsatisfied_predicates(qasm_crx_program: str):
         rebase(qasm_crx_program, gateset)
 
 
+# move to pyqasm
 def test_rebase_unsatisfied_predicates_no_check(qasm_crx_program: str):
     """Test that rebase raises an error if the predicates are not satisfied"""
     assert rebase(qasm_crx_program, {"h"}, require_predicates=False) == qasm_crx_program
 
 
+# move to pyqasm
 def test_rebase_over_supported_decomposition_basis_set(
     qasm_crx_program: str, qasm_crx_decomposed: str
 ):